Changed sReplyPortInUse type from long to int32, so it be used with atomic_add() also on 64 bit architectures.
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@42174 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ headers/os/app/Message.h
@@ -376,7 +376,7 @@ class BMessage {
 
 		enum			{ sNumReplyPorts = 3 };
 		static port_id	sReplyPorts[sNumReplyPorts];
-		static long		sReplyPortInUse[sNumReplyPorts];
+		static int32	sReplyPortInUse[sNumReplyPorts];
 		static int32	sGetCachedReplyPort();
 
 		static BBlockCache* sMsgCache;
